Move MemoryFailPoint to shared CoreLib partition.
